Applying mock manager as a global fixture. Integrating MockManager with mapper tests. All tests running successfully locally
