+++ requirements.txt
@@ -53,10 +53,13 @@ Protego==0.4.0
 ptyprocess==0.7.0
 pure_eval==0.2.3
 Pygments==2.19.1
-pytest==8.3.5
-pytest-asyncio==0.26.0
+pytest==7.4.3
 pytest-mock==3.14.0
-python-dateutil==2.9.0.post0
+
+# Utilities
+python-dotenv>=1.0.0
+
+pyyaml>=6.0.2
 PyYAML==6.0.2
 pyzmq==26.4.0
 redis==6.0.0
